Dashboard configuration

+++ app/Http/Controllers/ReservationController.php
@@ -11,6 +11,7 @@ class ReservationController extends Controller
     public function index()
     {
         $reservations = Reservation::with('room', 'room.hotel')
+            ->where('user_id', \Auth::user()->getUserInfo()['sub'])
             ->orderBy('arrival','asc')
             ->get();
 
@@ -26,7 +27,8 @@ class ReservationController extends Controller
 
     public function store(Request $request)
     {
-        $request->request->add(['user_id' => 1]);
+        $user_id = \Auth::user()->getUserInfo()['sub'];
+        $request->request->add(['user_id' => $user_id]);
         Reservation::create($request->all());
 
         return redirect('dashboard/reservations')->with('success','Reservation created!');
@@ -37,36 +39,55 @@ class ReservationController extends Controller
         $reservation = Reservation::with('room', 'room.hotel')
             ->get()
             ->find($reservation->id);
-        $hotel_id = $reservation->room->hotel_id;
-        $hotelInfo = Hotel::with('rooms')->get()->find($hotel_id);
 
-        return view('dashboard.reservationSingle', compact('reservation', 'hotelInfo'));
+        if ($reservation->user_id == \Auth::user()->getUserInfo()['sub']) {
+            $hotel_id = $reservation->room->hotel_id;
+            $hotelInfo = Hotel::with('rooms')->get()->find($hotel_id);
+
+            return view('dashboard.reservationSingle', compact('reservation', 'hotelInfo'));
+        }
+        else
+            return redirect('dashboard/reservations')->with('error', 'You are not authorized to see that.');
     }
 
-    public function edit(Reservation $reservation)
-    {
-        $reservation = Reservation::with('room','room.hotel')
+    public function edit(Reservation $reservation) {
+        $reservation = Reservation::with('room', 'room.hotel')
             ->get()
             ->find($reservation->id);
-        $hotel_id = $reservation->room->hotel_id;
-        $hotelInfo = Hotel::with('rooms')->get()->find($hotel_id);
 
-        return view('dashboard.reservationEdit', compact('reservation','hotelInfo'));
+        if ($reservation->user_id === \Auth::user()->getUserInfo()['sub']) {
+            $hotel_id = $reservation->room->hotel_id;
+            $hotelInfo = Hotel::with('rooms')->get()->find($hotel_id);
+
+            return view('dashboard.reservationEdit', compact('reservation', 'hotelInfo'));
+        } else
+            return redirect('dashboard/reservations')->with('error', 'You are not authorized to do that');
     }
 
-    public function update(Request $request, Reservation $reservation)
-    {
-        $reservation->user_id = 1;
+    public function update(Request $request, Reservation $reservation) {
+        if ($reservation->user_id != \Auth::user()->getUserInfo()['sub'])
+            return redirect('dashboard/reservations')->with('error', 'You are not authorized to update this reservation');
+
+        $user_id = \Auth::user()->getUserInfo()['sub'];
+        $reservation->user_id = $user_id;
+        $reservation->num_of_guests = $request->num_of_guests;
+        $reservation->arrival = $request->arrival;
+        $reservation->departure = $request->departure;
+        $reservation->room_id = $request->room_id;
+
         $reservation->save();
 
-        return redirect('dashboard/reservations')->with('success','Successfully updated your reservation!');
+        return redirect('dashboard/reservations')->with('success', 'Successfully updated your reservation!');
     }
 
-    public function destroy(Reservation $reservation)
-    {
+    public function destroy(Reservation $reservation) {
         $reservation = Reservation::find($reservation->id);
-        $reservation->delete();
 
-        return redirect('dashboard/reservations')->with('success','Successfully deleted your reservation!');
+        if ($reservation->user_id === \Auth::user()->getUserInfo()['sub']) {
+            $reservation->delete();
+
+            return redirect('dashboard/reservations')->with('success', 'Successfully deleted your reservation!');
+        } else
+            return redirect('dashboard/reservations')->with('error', 'You are not authorized to delete this reservation');
     }
 }
